feat:the mock file uploader is done

+++ media-service/media-biz/src/main/java/com/hasten/sim/Block.java
@@ -3,10 +3,15 @@ package com.hasten.sim;
 import lombok.Data;
 
 /**
+ * TODO:
+ *
+ * a field called `isDone` should added to tell the server that all blks have been uploaded.
+ *
+ *
  * @author Hasten
  */
 @Data
 public class Block {
     private final String fileName;
-    private final long blockId;
+    private final int blockId;
 }
